requirements.txt
Django==5.2.3
psycopg2-binary==2.9.10
djangorestframework==3.16.0
pytest==8.4.1
